M8-Full_stack/flask-blog-project-day1-connecting-with-python/requirements.txt
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
python-slugify==5.0.2
text-unidecode==1.3
